refactor plugins into multiple modules
